requirements.txt
# Core Django
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.9.0
gunicorn==21.2.0

# Environment & Configuration
python-dotenv==1.1.1

# HTTP & Networking
requests==2.32.3
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# Database (PostgreSQL for production)
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Static Files
whitenoise==6.6.0

# ML/AI for Recommendations
numpy==1.26.4
scikit-learn==1.4.0

# Other
asgiref==3.9.1
sqlparse==0.5.3
packaging==25.0